+++ com.devsquad10.shipping/src/main/java/com/devsquad10/shipping/application/dto/request/ShippingPostReqDto.java
@@ -1,4 +1,4 @@
-package com.devsquad10.shipping.application.dto;
+package com.devsquad10.shipping.application.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
