+++ com.devsquad10.shipping/src/main/java/com/devsquad10/shipping/application/dto/request/ShippingUpdateReqDto.java
@@ -1,6 +1,5 @@
-package com.devsquad10.shipping.application.dto;
+package com.devsquad10.shipping.application.dto.request;
 
-import java.io.Serializable;
 import java.util.UUID;
 
 import com.devsquad10.shipping.domain.enums.ShippingStatus;
@@ -11,14 +10,20 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ShippingResDto implements Serializable {
-
-	private UUID id;
+@Builder
+public class ShippingUpdateReqDto {
 
 	private ShippingStatus status;
 
 	private UUID orderId;
+
+	private String address;
+	private String requestDetails;
+
+	private String recipientName;
+	private String recipientPhone;
+
+	private UUID companyShippingManagerId;
 }
