+++ com.devsquad10.shipping/src/main/java/com/devsquad10/shipping/application/dto/response/ShippingResDto.java
@@ -1,5 +1,6 @@
-package com.devsquad10.shipping.application.dto;
+package com.devsquad10.shipping.application.dto.response;
 
+import java.io.Serializable;
 import java.util.UUID;
 
 import com.devsquad10.shipping.domain.enums.ShippingStatus;
@@ -10,20 +11,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
-public class ShippingUpdateReqDto {
+public class ShippingResDto implements Serializable {
+
+	private UUID id;
 
 	private ShippingStatus status;
 
 	private UUID orderId;
-
-	private String address;
-	private String requestDetails;
-
-	private String recipientName;
-	private String recipientPhone;
-
-	private UUID companyShippingManagerId;
 }
