+++ com.devsquad10.shipping/src/main/java/com/devsquad10/shipping/application/exception/shipping/ShippingNotFoundException.java
@@ -1,4 +1,4 @@
-package com.devsquad10.shipping.application.exception;
+package com.devsquad10.shipping.application.exception.shipping;
 
 public class ShippingNotFoundException extends RuntimeException {
