+++ com.devsquad10.shipping/src/main/java/com/devsquad10/shipping/application/exception/shippingAgent/ShippingAgentTypeNotFoundException.java
@@ -1,4 +1,4 @@
-package com.devsquad10.shipping.application.exception;
+package com.devsquad10.shipping.application.exception.shippingAgent;
 
 public class ShippingAgentTypeNotFoundException extends RuntimeException {
 	public ShippingAgentTypeNotFoundException(String message) {
